YOU DON'T HAVE TO EXIT OUT EVERY TIME YOU WANNA START A NEW ROUND NOW!
Added HowToPlayManager script for button management on the HowTo (Test) scene.
Added new tags to the tag manager (PlayerManager and LevelSelectManager) and attacked them to the appropriate objects (the PlayerManager found on PlayerSetup scene and the LevelSelectManager found on the LevelSelect scene).
Edited GameManager and LevelSelectManager scripts to destroy appropriate objects when needed.
Edited PlayerSetup scene with text outlining to press a button to recognize a controller.
Edited buttons on LevelSelect scene because they work now!
Edited buttons on HowTo (Test) because they also work now!
Added a button to EndingScene to bring players back to LevelSelect.